chore: deprecate `composio integrations add`
